[Maintenance] Reorganize converters' namespaces and provide backward compatibility
