+++ spec/Converter/LineItem/ShipmentLineItemsConverterSpec.php
@@ -11,14 +11,14 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\RefundPlugin\Converter;
+namespace spec\Sylius\RefundPlugin\Converter\LineItem;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Core\Model\AdjustmentInterface;
 use Sylius\Component\Core\Model\ShipmentInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
-use Sylius\RefundPlugin\Converter\LineItemsConverterInterface;
+use Sylius\RefundPlugin\Converter\LineItem\LineItemsConverterInterface;
 use Sylius\RefundPlugin\Entity\LineItem;
 use Sylius\RefundPlugin\Exception\MoreThanOneTaxAdjustment;
 use Sylius\RefundPlugin\Model\OrderItemUnitRefund;
