+++ src/Converter/LineItem/LineItemsConverterInterface.php
@@ -1,0 +1,27 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\RefundPlugin\Converter\LineItem;
+
+use Sylius\RefundPlugin\Entity\LineItemInterface;
+use Sylius\RefundPlugin\Model\UnitRefundInterface;
+
+interface LineItemsConverterInterface
+{
+    /**
+     * @param UnitRefundInterface[] $units
+     *
+     * @return LineItemInterface[]
+     */
+    public function convert(array $units): array;
+}
