+++ src/Converter/RequestToOrderItemUnitRefundConverter.php
@@ -13,23 +13,11 @@ declare(strict_types=1);
 
 namespace Sylius\RefundPlugin\Converter;
 
-use Sylius\RefundPlugin\Model\OrderItemUnitRefund;
-use Symfony\Component\HttpFoundation\Request;
-
-final class RequestToOrderItemUnitRefundConverter implements RequestToRefundUnitsConverterInterface
-{
-    public function __construct(private RefundUnitsConverterInterface $refundUnitsConverter)
-    {
-    }
-
+if (false) {
     /**
-     * @return OrderItemUnitRefund[]
+     * @deprecated since 1.4, to be removed in 2.0, use Sylius\RefundPlugin\Converter\Request\RequestToOrderItemUnitRefundConverter instead
      */
-    public function convert(Request $request): array
+    final class RequestToOrderItemUnitRefundConverter extends Request\RequestToOrderItemUnitRefundConverter
     {
-        return $this->refundUnitsConverter->convert(
-            $request->request->all()['sylius_refund_units'] ?? [],
-            OrderItemUnitRefund::class,
-        );
     }
 }
